+++ ImagingKit_Filter/src/main/java/hageldave/imagingkit/filter/implementations/ConvolutionFilter.java
@@ -1,4 +1,4 @@
-package hageldave.imagingkit.filter.filterimpl;
+package hageldave.imagingkit.filter.implementations;
 
 import java.util.function.Consumer;
