+++ ImagingKit_Filter/src/main/java/hageldave/imagingkit/filter/implementations/GaussFilter.java
@@ -1,4 +1,4 @@
-package hageldave.imagingkit.filter.filterimpl;
+package hageldave.imagingkit.filter.implementations;
 
 public class GaussFilter extends ConvolutionFilter {
